requirements.txt
Django==2.0
django-filter==1.0.4
djangorestframework==3.7.3
django-rest-swagger==2.1.2
gunicorn==19.7.1
psycopg2==2.7.1
django-storages==1.6.3
boto==2.48.0
ipython
raven
PyYaml
coreapi

# Testing Utilities
faker
responses==0.8.1
ipdb
coverage
codacy-coverage
django-cors-headers
django_nose